Implemented validation of template XML in ImportCSV (closes #149). Minor bug fixes in soap_server.template.assign_type_to_resource.
git-svn-id: svn+ssh://ec2-54-229-95-247.eu-west-1.compute.amazonaws.com/home/svn/svnrepository/HYDRA@303 5b2e1b5e-a1f6-4386-8dd1-2f88c1042e8b

+++ HydraServer/trunk/python/soap_server/hydra_complexmodels.py
@@ -3,6 +3,7 @@ from spyne.model.primitive import Unicode
 from spyne.model.primitive import String
 from spyne.model.primitive import Integer
 from spyne.model.primitive import Decimal
+from spyne.model.primitive import Float
 from spyne.model.primitive import DateTime
 from spyne.model.primitive import AnyDict
 from spyne.model.primitive import Double
@@ -227,7 +228,7 @@ class EqTimeSeries(HydraComplexModel):
 
 class Scalar(HydraComplexModel):
     _type_info = [
-        ('param_value', Decimal),
+        ('param_value', Float),
     ]
     def __init__(self, val=None):
         super(Scalar, self).__init__()
